requirements.txt
fastapi
uvicorn
gmqtt
redis
websockets
python-dotenv